Travis: Refactor out lfs_file_info
